Add SQLite database initialization and connect site
